Extract common typed matchers logic
